+++ farm-staking/farm-staking-proxy/src/proxy_actions/mod.rs
@@ -1,0 +1,4 @@
+pub mod claim;
+pub mod merge_pos;
+pub mod stake;
+pub mod unstake;
